+++ tools/xdhwebq/xdwrgstry.cpp
@@ -22,7 +22,7 @@
 using namespace xdwrgstry;
 
 entry___ xdwrgstry::parameter::ModuleFilename( "ModuleFilename", sclrgstry::Parameters );
-entry___ xdwrgstry::parameter::script::Fallback( "Fallback", xdwrgstry::Scripts );
+